+++ docs/source/en/api/models/asymmetricautoencoderkl.mdx
@@ -44,7 +44,7 @@ image.save("image.jpeg")
 
 ## AsymmetricAutoencoderKL
 
-[[autodoc]] AsymmetricAutoencoderKL
+[[autodoc]] models.autoencoder_asym_kl.AsymmetricAutoencoderKL
 
 ## AutoencoderKLOutput
