+++ src/diffusers/models/__init__.py
@@ -16,7 +16,8 @@ from ..utils import is_flax_available, is_torch_available
 
 
 if is_torch_available():
-    from .autoencoder_kl import AsymmetricAutoencoderKL, AutoencoderKL
+    from .autoencoder_asym_kl import AsymmetricAutoencoderKL
+    from .autoencoder_kl import AutoencoderKL
     from .controlnet import ControlNetModel
     from .dual_transformer_2d import DualTransformer2DModel
     from .modeling_utils import ModelMixin
